clause_exp/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.0
qdrant-client==1.7.1
sentence-transformers==2.2.2
openai==1.3.7
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0
structlog==23.2.0
tenacity==8.2.3
numpy==1.24.3
